ci: setup renovate for wolfi packages
Adding the new get package action in renovate workflow + mount the generated files to /tmp/renovate/wolfi-packages

Renovate is not able to access directories outside of the repository directory it creates when using renovate action.
Since repository directory is recreated everytime renovate starts, we definitely can't pass the wolfi packages files to it.
Instead we'll serve the mounted directory as an API inside the renovate container, and set the custom datasource to query localhost.

Also added a group rule for grouping wolfi packages on the same PR, and a filter on iptables & ip6tables as we want to control those upgrades manually.

Signed-off-by: Antony Reynaud <antony.reynaud@isovalent.com>