fqdnha: stop waiting for endpoint regen before opening socket
We were waiting for all endpoints to regenerate before opening the
socket. We did this because, previously, the relay would not have access
to the full set of rules until the DoubleProxy had started initializing
(and did the dance with the singleton Proxy).

This is no longer necessary; the DoubleProxy now is a decorator and thus
never misses rules. Likewise, the only gRPC methods implemented are
SubscribeFQDNRules and NotifyOnDNSMsg. Both of these function correctly
as the agent is initializing, so we no longer need to delay listening.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>