enterprise/fqdn-proxy: various cleanups
Make proxyContext mutex a value instead of a pointer to simplify
initialization. There is no point in keeping it behind a pointer because
proxyContext shouldn't be copied anyway. Also rename to mu while at it
to reduce stutter and move it above the ipCacheV1 member it protects.

Also use slices.Contains to simplify (*SimpleSelector).Selects.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>