fqdnha: implement agent-side state synchronization
This adds the server implementation of SubscribeState. It creates two
new statedb tables, one each for the agent and proxy's state. It
implements a simple bidi streaming server that pushes agent state to the
proxy, and pushes proxy state to the statedb.

Finally, it adds a simple test that ensures state transitions are
captured as expected.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>