egwha: only upsert matched endpoints on change.
When doing and policy endpoint matching for all policies
we currently always recompute the matched endpoints set
from scratch and then re-insert this back into the config
table.

Previously, when we stored policyConfigs in a map there
wasn't much cost to doing this, however inserts are
more expensive in statedb as they often require copying
of internal radix-tree nodes as well as potentially
upsizing/downsizing nodes.

We can likely avoid a lot of this work if we keep track
of whether the set has changed since the last update.

Thus, in such case that the set of matched endpoints for
a policy doesn't change we can avoid doing the update.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>