fqdnha: implement proxy-side state synchronization
Similar to the previous commit, this implements the client side of
SubscribeState. It initializes the two same statedb tables, and adds a
simple method for synchronizing state changes between the two processes.

It also adds the necessary edge detection within the proxy's
notification queue, so that we can correctly transition between the
states:

- LIVE: all is working as planned
- AGENT_OFFLINE: messages are being queued
- REPLAYING: connected to agent, but there is a backlog of DNS messages
- WAITING_FOR_AGENT_LIVE: replay is complete, but agent has not finished
  regenerating

These states will be used to mediate the transition out of offline mode.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>