ilb: enable masquerading
For the Standalone ILB case, masquerading traffic wasn't necessary,
therefore the recommendation was to disable it completely.

ILB's CNP integration changes the source address of an incoming connection
to the one of the node-local Ingress IP, to be able to later match it
against `CiliumNetworkPolicies` that use the reserved Ingress identity.
This source address replacement is done for all incoming connections, no
matter their destination address.

This leads to issues where cluster-external backends (e.g. `LBBackendPools`
with backend type `IP` or `HostName`) don't know how to route the
Ingress IP.

Therefore, enabling ILB's CNP integration also requires that traffic
leaving the cluster needs to be masqueraded with the Node IP - but
without masquerading the traffic targeting cluster-internal Pods.
Otherwise CNP using the reserved Ingress identity would no longer work
as expected because the policy enforcement sees the Node IP as source
address (instead of the expected Ingress IP).

Therefore it's also required to configure the native routing CIDR
(that should be excluded from being masqueraded) correctly per IP
family. For development and testing (CI) purposes, we set the routing
to the Pod CIDR. Customers need to adjust this accordingly.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>