fqdnha: agent: pause regeneration until remote proxy replay complete
When offline mode is enabled, the remote proxy will queue DNS messages
and write them to the outgoing BPF ipcache. When the agent is back up,
we need to learn about those new IPs before we proceed with
regeneration. Otherwise, packets may drop as endpoints are regenerated
over to the "new" bpf ipcache.

The timeline is thus:

1. Agent goes down
2. Proxy learns about a new IP. It writes this to the old IPCache
3. Agent comes back up.
4. The queue of new IPs is replayed to the agent
5. When the queue hits zero, regeneration can proceed

This ensures that all IPs are in both old and new ipcaches, and no
traffic will be dropped.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>