fqdnha: delay opening DNS proxy when running in offline mode
When offline mode is enabled, we need to ensure that all DNS requests go
through the remote proxy until all endpoints are regenerated.

This is because any DNS request may include a new IP. We need to program
that IP in to the bpf ipcache, but any un-regenerated endpoints are
still referencing the old ipcache. The remote proxy, however, can still
write to it!

So, all DNS requests must go to the remote proxy, which will write to
the old IPCache and forward it to the agent, which will write to the new
ipcache.

Once all endpoints are regenerated, we can start the local proxy.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>