ilb: disable IPv6 on JWKS clusters
Currently, we configure all ILB clusters to use IPv4 only - except for
the Envoy cluster that is used for fetching the remote JWKS.

To avoid confusion and potential issues, this commit configures the
JWKS cluster to only use IPv4 too.

Note: Once we have IPv6 added for ILB we should probably configure the
IP families on the ILB related Envoy clusters to match the configured
IP families in the Cilium Helm Chart.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>