ent/renovate: group enterprise go deps updates in all branches
Currently, renovate is configured to only group go dependency updates
in the main-ce branch. This is most likely a leftover from a similar
configuration used in cilium/cilium, as there we don't want to
automatically bump dependencies in stable branches there. Given the
same policy doesn't seem to be applied here, let's make sure to group
the updates in all branches, to avoid a proliferation of renovate PRs
in stable branches, which add a lot of burden for reviews/triaging.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>