enterprise/fqdn-proxy: add remoteNameManager store for selector updates
Previous commits have introduced streams of FQDN identities and
selectors from the agent to the fqdn-proxy. Now add the
remoteNameManager and the plumbing which inserts the streamed identities
and selectors into its stores, with all the necessary conversions from
protobuf messages.

Avoid storing identities which are not of interest to the proxy.
Identities formed of labels which are not exclusively of source FQDN
(with the exception of the world label) can't be handled as of now,
hence we discard them here. Additionally, a future commit will remove
the streaming of these identities to the proxy in the first place.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>