enterprise/fqdn-proxy: look up identity via remoteNameManager
Let the remoteNameManager keep the reference to the BPF ipcache
interface. This also allows to easily mock out BPF ipcache map access
for tests.

The IPCache version is no longer provided by the agent in the status
message, so supportsIPCacheV1 can never become true. Instead, wire the
lookup thourgh remoteNameManager which knows whether BPF ipcache lookup
is allowed or not.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>