egwha: reduce copying of PolicyConfig fields.
Since introducing statedb as the primary way of storing
egwha policyConfig state we've seen a modest increase
in memory usage.

One of the reasons for this is that we need to do object
clone on the policyConfigs we insert into the statedb
table.
Statedb depends on unique references and will error if
the same object pointer is inserted into the database.
Statedb depends on this property to track unique objects
across various operations.

However, one property of the AgentPolicyConfig is that
it's components are always treated as immutable:

* PolicyConfig: base shared with operator, models a iegp
	policy at runtime: this is always parsed and recreated
	from scratch upon iegp policy events.
	The Agent control plane only ever reads this data and
	never mutates it following it being created.

* gatewayConfig: local gateway configuration derived from
	PolicyConfig: Always derived from PolicyConfig and
	(re)created from scratch before overwriting the
	old gatewayConfig.

* matchedEndpoints: Set of known endpoints matched to this
	policy (given the policy label matcher). This is
	regenerated often following endpoint updates but
	is always recomputed from scratch.

Therefore, we can just clone the top level struct and not
worry about shared references so long as we maintain this
property of accessing the data.

Note: We could assure this going forward by defining opaque
interface types for the parts of the AgentPolicyConfig,
however as this code is not used by other components besides
EGWHA we can keep things simple for now and just convert
the various fields to pointer types to reduce memory copying.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>