ilb: configure quarantime timeout on T1 services correctly
Currently, ILB controlplane configures the T1 k8s services with the
annotation `service.cilium.io/health-check-quarantine-timeout: 0s`.
Following the description of the annotation, this was ok, because this
was only meant as timeout to wait before resuming health checking in
case of quarantining the service.

```
// ServiceHealthQuarantineTimeout annotation determines the timeout duration
// for a given backend to reside in unhealthy state before probes are resumed
// again.
// Allowed values:
//  - A duration, for example:
//    "service.cilium.io/health-check-quarantine-timeout": "30s"
ServiceHealthQuarantineTimeout = ossannotation.ServicePrefix + "/health-check-quarantine-timeout"
```

It looks like the new healthchecker implementation for the new Cilium LB
controlplane interprets this config option slightly different. It interprets
the value as general probe interval during the state "unhealthy".

See https://github.com/isovalent/cilium/blob/main-ce/enterprise/pkg/service/healthchecker/tables.go#L71-L81

This results in a high amount of probes if the quarantine timeout
is set to `0s`.

Therefore, this commit is adjusting the value to be the same as the
general probe interval.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>