ilb: ci: write T2 Envoy accesslog to file and store it
By default, the T2 Envoy accesslog is written to stdout of the Cilium
Proxy pod. In case of an error, the Envoy logs from the sysdump (containing
the Envoy application & accesslog) don't hold that much information due to
the amount of applicationlogs written.

Therefore this commit changes CI to write the Envoy accesslog to a file
that gets fetched and uploaded as GitHub action artifact in case of an error.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>