ent/maps: generalize the external endpoints policy map
Currently, the external endpoints policy map of maps is strictly tied
to Cilium Mesh. However, a very similar approach is being adopted in
the context of the private networking effort, and it is helpful being
able to reuse the already existing map infrastructure.

Hence, this commit generalizes the map creation and configuration logic
to be independent from Cilium Mesh itself. At a very high level, the
[Writer] interface, which is provided via Hive, allows to upsert and
delete entries. If enabled, the map is created automatically via the
Start hook, and similarly GCed (if present) when not enabled.

Finally, on the BPF side, the definition of the map itself, and of the
helper functions, are moved to a separate file, again for easier reuse.
No additional changes are applied here, besides generalizing the names,
and minor adaptations to please checkpatch.

The main differences compared to the previous implementation are:

1. The name of the map is changed, but the upgrade path is not handled
   (i.e., the new map is re-created from scratch, and the old one is
   left behind). It would be possible to improve this, but it looks
   unnecessary considering the overall state of the feature, and the
   fact that Cilium Mesh is not used anywhere to the best of my knowledge.
2. Map creation is no longer blocked on host datapath initialization.
   It is honestly unclear to me why that was necessary (the original
   commit [1] does not provide enough information), but no other map
   adopts that approach. Additionally, that goes against the idea that
   maps are always created (and initialized, in many cases) before
   endpoints regeneration. I wonder if that was necessary due to some
   mismatch between the map definition on the datapath and agent sides.
   Edit: more context available in [2]. I haven't hit that issue for
   the moment --- let's start with the easier and more natural approach,
   we can always reevaluate later if it pops up in specific environments.
3. The size of the outer map is now configurable via a dedicated flag,
   and the inner one matches that of the policy maps.

[1]: c46e50c1b5d9 ("mesh: wait on datapath before opening the cilium mesh policy map")
[2]: https://isovalent.slack.com/archives/C01G1Q0AVC4/p1708414697880329

Co-authored-by: Harsimran Pabla <hpabla@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>