ent/maps: implement ext-eps policy map writer via statedb table
Rather than directly performing the operations against the BPF map, let's
add an intermediate statedb table and make use of the generic reconciler
logic, so that is automatically takes care of retries in case of errors,
as well as of pruning of stale entries upon agent restart. To this end,
the [Writer] interface is extended to allow providing an initialization
signal, so that stale entries are not pruned too early.

While being there, let's also implement a (privileged) unit test to
validate the current functioning of this module.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>