ent/bpf: propagate error in case of external_eps map miss
Currently, the __external_eps_policy_can_access helper ignores possible
misses in the external endpoints policy map, and simply allows the
packet through returning CTX_ACT_OK. Hence, it is basically failing
open, which is definitely problematic from a security point of view.

In preparation for the reuse of the same helper in the context of
private networking, where we want to adopt a stricter approach, let's
make the function return an error to the caller, so that it is possible
to distinguish this case. The existing cilium mesh logic is adapted
to simply ignore this error to preserve the existing behavior, as it
would otherwise completely break connectivity without further changes,
which are out of scope here.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>