gha: fix execution of scheduled workflows on stable branches
[1] recently changed the workflow upstream, moving the `v` branch
prefix directly to the branch name definition. However, the same
change got missed when synchronizing that commit [2], hence causing
this workflow to consistently fail.

Additionally, let's change the `hourModulo` for the v1.18-ce branch,
as it currently corresponds to 0 when taking the modulo, and a comment
there warns not to use 0.

[1]: 3abf4c111c23 ("Prepare for v1.19 development cycle")
[2]: 71ffa4436aac ("Merge commit '57f801a7c1ed84df875bddfa1265b369ddf5aeea' into pr/raszabo/main-sync-2025-08-05")

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>