ilb: add multi-namespace test with in-cluster hostname
Add a test that makes requests of LBServices in multiple namespaces. The
LBServices are reached via LBBackendPools that use the in-cluster
service hostname. This is the first test to use the in-cluster
service hostname.

Refactor the test scenario data types and helper functions to support
multiple namespaces. The namespaces are created and cleaned up for each
test run.

The tests can be invoked as follows:

    cd cilium-cli
    make
    ./cilium lb test --run 'TestHTTPMulti.*' --verbose

To keep the resources intact after the test has been run:

    ./cilium lb test --run 'TestHTTPMulti.*' --verbose --cleanup=false

Signed-off-by: Andrew Sauber <andrew.sauber@isovalent.com>