ilb: add a multi-namespace test for in-cluster services
Adds a test which reaches LBServices in multiple namespaces. The
LBBackendPools references the services using K8sServiceRef (rather than
in-cluster hostname).

The test deploys services in three namespaces, which can easily be
extended to any number of namespaces by adding a list of names to the
namespaces slice.

To view the access logs that are generated by this test, you can add
the following helm configuration to `enterprise-kind-loadbalancer.yaml`
when running the local kind environment:

    loadbalancer-cp-accesslog-enable-tcp: "true"
    loadbalancer-cp-accesslog-enable-udp: "true"

These can be added at the root of the values under `extraConfig.`

The resulting access logs lines then start with a per-namespace field in
the default configuration, which look like the following:

    [2025-08-11T10:50:33.863Z][access][tcp] service.namespace="backend-2"
    ...
    [2025-08-11T10:50:33.875Z][access][healthcheck] service.namespace="backend-3"
    ...
    [2025-08-11T10:50:33.893Z][access][healthcheck] service.namespace="backend-3"

This can be observed with

    k logs -f -n kube-system cilium-envoy-24qpf | grep '\[access\]'

using any of the envoy pod names which received traffic during the test.

Signed-off-by: Andrew Sauber <andrew.sauber@isovalent.com>