ilb: Add a version guard for TestHTTPJWTAuth claim based filtering
JWT claim based filtering is only supported after v1.18-ce, but
currently, we're missing the version guard for this test case. This
breaks the CI in the v1.17-ce because v1.17-ce is not using in-tree
cilium-cli for testing and still relying on the latest released version.
Fix this by introducing a version guard.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>
Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>