privnet: introduce the configuration flags
Introduce the private networks configuration, to allow enabling the
feature, as well as configuring its operational mode, between default
(i.e., main cluster) and bridge (i.e., interconnection with external
endpoints). The usage of "mode", instead of a simple boolean, is
intended to simplify adding new operational modes in the future,
if necessary.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>