ilb: prevent load on api server due to node change events
ILB controlplane watches Node resources via `NodeSource` that listenes for changes
on the Node resource and translates them into controller-runtime events. Currently,
every Resource event is forwarded as controller-runtime event.

Due to the fact that Node resources can change quiet frequently (~5s), this can trigger
a lot of `LBService` reconciliations which eventually can lead to a lot of load on the
API server (trying to update T2 `CiliumEnvoyConfig` & T1 K8s `Service`).

All of this isn't really necessary because the ILB controlplane is actually only
interested in the labels and addresses of a Node. Therefore, this commit modifies the
`ciliumNodeSource` to keep track of the addresses per Node and only forwards the events as
controller-runtime events in case of any changes - which is either on a new or deleted Node
or when an address/label of an existing Node changes.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>