privnet: introduce the kvstore Endpoint type
The Endpoint type represents a PrivateNetworkEndpointSlice entry,
and it is intended to be stored in etcd to transfer the information
cross-cluster via clustermesh. Similarly to `state/ip/v1` entries,
each entry represents a single IP address (either IPv4 or IPv6).

Entries are keyed by source cluster, network name and endpoint IP
address (in pod network). The reason for including the IP as part
of the key is that none of the other properties alone guarantees
uniqueness, in case e.g., multiple bridge nodes expose access to
the same entries (the pod IP would be unique there as well). To
be fair, the network name does not contribute to uniqueness, but
it is still added as part of the key to make it more user-friendly,
and allow easily filter all entries belonging to the same network.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>