cilium: Split out config control to enable IPIP devices
[ upstream commit 83a4bc36a975c2928507e033dc4a4d6723dbcfcc ]

Currently the IPIP device gets enabled via two different options, that is,
i) when IPIP termination gets enabled, or ii) when the socket health data
path is opted-in. As we are adding a third option, namely when the IPIP
DSR dispatch is configured for non-XDP, lets add an intermediary config
knob so we can simplify the loader part and don't have to check for all
three conditionals each time. Also, add a fallback to index 0 (invalid
ifindex) in case this was not defined (e.g. in case of bpf_lxc which is
tcx when XDP acceleration is used on physical device and the nodeport
code gets included but not used from bpf_lxc).

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>