privnet: introduce IDPool for private networks ID allocation
Introduce a simple logic to handle the allocation of IDs for private
networks, to be used as identifiers on the datapath side. The pool
leases IDs sequentially, keeping track of the ones subsequently
released. The tracking logic is not particularly optimized, given
that relatively few private networks are expected to exist at any
given point in time.

Given that IDs are allocated independently by each agent, and they
are not stable across restarts, the pool is configured to use a
random initial ID, to avoid incorrectly assuming that IDs are stable
when testing with only one or two private networks. Still, it is
possible to override the pool in testing environments, to be able
to have stable IDs there.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>