bpf: Fix upgrade to lb algorithm selection per annotation
[ upstream commit 26b5bbc464e63d966040e3904e1594a7821eb884 ]

Dmitriy says:

  When --bpf-lb-algorithm-annotation feature is enabled, BPF code
  in run-time may not select load balancer algorithm for service at
  all for previously existing services. Thus, this behavior broke
  all new network connections to the service if it had an unknown
  algorithm value in service bpf map entry. This bug is stably
  reproducible for NodePort, HostPort, LocalRedirect service types.

  This commit solves the problem as follows: in the situation where
  the BPF service map contains an unknown LB algorithm, it simply uses
  the default LB algorithm from the --bpf-lb-algorithm option.

lb{4,6}_algorithm() should return a proper algorithm that is going to
be used in the datapath. Either a service had an annotation, or if not
then the user configured default should be picked.

One corner case is when we come from a future Cilium version where
users had an algorithm annotation on the service, which the current
Cilium version does not support. In that case we can only treat this
as a hint and need to fallback to the default.

Note that in the old code before the LB control plane rework, the
GetAnnotationServiceLoadBalancingAlgorithm() was pushing through
any annotation which was not loadbalancer.SVCLoadBalancingAlgorithmUndef
and otherwise the function was returning the default selected algorithm.
After the rework we just propagate loadbalancer.ToSVCLoadBalancingAlgorithm()
directly.

This meant that handling of loadbalancer.SVCLoadBalancingAlgorithmUndef
was pushed to runtime, therefore for services with no explicit annotation
this triggered the default case which led to drops.

Another side-note: loadbalancer.ToSVCLoadBalancingAlgorithm() does not
translate LB_SELECTION_FIRST. The latter is only ever used in BPF unit
tests.

Co-developed-by: Dmitriy Andreychenko <dmitriy.andreychenko@flant.com>
Signed-off-by: Dmitriy Andreychenko <dmitriy.andreychenko@flant.com>
Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>