privnet: introduce clustermesh observer for endpoint events
EndpointsObserver implements the [store.Observer] interface so that it
can be hooked into the clustermesh logic to watch privneteps entries,
and exposes an observable interface for consumption by the appropriate
reconciler. Observed events are queued for up to 50 milliseconds to
allow batching a set of events into a single transaction when writing
them to the appropriate table.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>