privnet: introduce the privnet-mapentries statedb table
Introduce the definition of the privnet-mapentries statedb table, which
is designed to store the subset of endpoint information, as well as
additional route-related entries to be eventually propagated into the
private networks related BPF maps.

Additionally, extend the [tables.Endpoint] type with helper functions
to convert it into a MapEntry, as well as create the corresponding
MapEntryKey.

Co-authored-by: David Bimmler <david.bimmler@isovalent.com>
Co-authored-by: Fabian Fischer <fabian.fischer@isovalent.com>
Co-authored-by: Harsimran Pabla <hpabla@isovalent.com>
Co-authored-by: Sebastian Wicki <sebastian@isovalent.com>
Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>