privnet: introduce the privnet-routes statedb table
This adds a new routes table to the private network package. The routes
table is intentionally kept simple for now and only contains the routes
present in the  privet network CR. Its main purpose is to allow
downstream consumers to query if a certain route exists within a private
network.

Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>