ci: configure repository & release-version in cilium-cli action
Currently, preparing a release branch requires to not only set
the Cilium CLI version but also change all enterprise GH workflows
to use the released cilium-cli version instead
of using the in-tree version (which only exists on `main-ce`). This
is done by setting `repository` & `release-version` properties on the
respective GH action.

Therefore, this commit defines `repository` & `release-version`
for on all usages of the `cilium-cli` GH action. This doesn't affect the
workflows on `main-ce` (because the used variable `CILIUM_CLI_VERSION` is empty -
https://github.com/isovalent/cilium/blob/main-ce/.github/actions/set-env-variables/enterprise-overrides.sh#L26-L27 &
https://github.com/cilium/cilium-cli/blob/main/action.yaml#L60-L71),
but eases the change when preparing a release branch.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>