egressgwha: add support for alternate interface name
Allow users to identify the desired egress interface in the IEGP with an
alternate interface name. This actually works quite well already, as
LinkByName() returns the correct link object when using the alternate
interface name.

What *doesn't* work is using the alternate name to access the `rp_filter`
control in /proc. Here we still need to use the primary interface name.
But luckily that's just an innocent change in how we process the
`egressGroups` - by using the effective interface name from the link
object, all subsequent code operates on the primary name. The fact that the
IEGP potentially specified an alternate name is hidden inside
deriveFromGroupConfig().

For consistency this also changes the parameter to
GetIfaceFirstIPv4Address(), but it shouldn't make a difference.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>