ilb: fix `cilium lb status` BGP peers 0/0
If the name of an `LBService` doesn't match the one
of the referenced `LBVIP` the command `cilium lb status` calculates
a wrong BGP peer status of `0/0`.

A usecase is the ILB E2E test where a VIP is shared between multiple lbservices.

This commit fixes the label matcher to use the referenced name of the
LBVIP on the `LBService` instead of its name.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>