ilb: fix `cilium lb status` BGP peer with multiple clients
Currently `cilium lb status` doesn't properly calculate the total
of BGP peers when a LBService is advertised to multiple clients - e.g.
BGP Peers `4/2`.

The reason is that the total of BGP peers only respects the number of
`BGPPeerConfigs` when calculating the total of BGP peers. But it needs to
use the actual peers by looking them up in the `BGPClusterConfigs` (plus
respecting the `BGPPeerConfigs`).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>