ipam: Register k8s sync metrics for cluster-pool
[ upstream commit 21331c8b38cf90782512295ce6ba39639b45bbab ]

[ Backporter's notes: Used global operator registry instead of the hive
provided one since it is not available in the stable branch. ]

In cluster-pool IPAM mode the allocator creates a metrics set related to
the activity of the internal trigger. Specifically, the trigger is
called (and the related metrics are updated) each time the internal
status of the CiliumNode, including the allocated pod CIDRs, is synced
to k8s.  There is however an issue, since the metrics are not correctly
registered to the metrics registry, therefore they are not actually
available.

This commit registers the metrics so that they can be correctly exposed
and visible with the `cilium-operator-generic metrics list` command.

Signed-off-by: Fabio Falzoi <fabio.falzoi@isovalent.com>