ilb: speed up fetching loadbalancer status model
Currently, the BGP peerconfigs & advertisements are fetched once for
every `LBService` when calculating the ILB status model that is used
not only by `cilium lb status` but also the operator when fetching the
metrics.

This slows down this logic a lot.

Therefore, this commit refactors the logic to fetch the resources
once and passes them to the logic.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>