privnet: improve decoupling between kvstore and statedb endpoint types
Currently, the statedb privnet endpoint type builds on top of the
kvstore one, for convenience and to avoid unnecessary code duplication.
However, it is a somewhat suboptimal that other packages, e.g., the
one implementing the reconcilers, started depending on types and
helpers defined in the kvstore package.

Let's improve decoupling by defining appropriate aliases and wrapper
functions as part of the tables package, so that we can still benefit
from code reuse, while keeping the packages more logically separate
and simplifying future modifications that may require extending the
types for internal use.

Reported-by: Fabian Fischer <fabian.fischer@isovalent.com>
Reported-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>