ci: svic tags for aks instances
In this commit we tag the created aks instance with svic_falco_supported
Tag is set to no because instances have a lifecycle less than 24h.

Nodegroups and nodes are tagged with svic_osquery_supported to no for
the same reason.

Updated the mindfulness script to add the AKS action as an exception.

Signed-off-by: Antony Reynaud <antony.reynaud@isovalent.com>