privnet: Add types package
This commit adds a types package to pkg/privnet. This types package is
intended to be imported by different packages and is designed to contain
a small set of dependencies. It currently defines the pod annotation
name and schema used by PrivNet. These types will also be imported by
various packages later.

Co-authored-by: David Bimmler <david.bimmler@isovalent.com>
Co-authored-by: Fabian Fischer <fabian.fischer@isovalent.com>
Co-authored-by: Harsimran Pabla <hpabla@isovalent.com>
Co-authored-by: Marco Iorio <marco.iorio@isovalent.com>
Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>