fqdnha: proxy: installation: pass offline mode
We have the helm value, we just need to pass it to the proxy via the
command line. This also renames the value from `enableOfflineMode`
to the more flexible `offlineMode.enabled`.

(This removes the unused value dnsRejectResponseCode, which was present
in values.yaml.tmpl but not values.yaml nor used in the templates).

Signed-off-by: Casey Callendrello <cdc@isovalent.com>