privnet: Add mock implementations for the endpoint infrastructure
This commit prepares the privnet package for better script testing. This
is achieved by defining a set of abstract endpoint interfaces and
implementing them two times:

The first implementation in adapters.go implements these interfaces on
to of the real endpoint manager and endpoint api.

The second implementation in mock_endpoints.go implements these
interfaces as a mock implementation to be used in script tests. The mock
implementation also simulates the endpoint restoration fences and allows
interaction with the mock endpoint manager via script commands.

The mock implementation is injected in tests with `cell.DecorateAll`,
this will cause Hive to ignore the implementation in adapters.go.

As part of this restructuring, the reconcilers package also gets its
own cell.

Co-authored-by: David Bimmler <david.bimmler@isovalent.com>
Co-authored-by: Fabian Fischer <fabian.fischer@isovalent.com>
Co-authored-by: Harsimran Pabla <hpabla@isovalent.com>
Co-authored-by: Marco Iorio <marco.iorio@isovalent.com>
Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>