privnet: Add pods reconciler
This commit adds a small reconciler which watches the local pod table
for privnet annotation changes and updates the active/inactive state of
the corresponding endpoint accordingly.

Because other reconcilers will have to be informed about changes to
the activation state of a privnet-enabled endpoint, this commit also
adds the `EndpointActivationManager` type, which will allow other
reconcilers to be informed whenever the `activatedAt` endpoint property
has changed.

The logic added in this commit is part of the script test added in the
subsequent commit.

Co-authored-by: David Bimmler <david.bimmler@isovalent.com>
Co-authored-by: Fabian Fischer <fabian.fischer@isovalent.com>
Co-authored-by: Harsimran Pabla <hpabla@isovalent.com>
Co-authored-by: Marco Iorio <marco.iorio@isovalent.com>
Co-authored-by: Tobias Klauser <tobias@isovalent.com>
Signed-off-by: Sebastian Wicki <sebastian@isovalent.com>