privnet: Intrduce VNI field to the ClusterPrivateNetwork resource
Introduce `.spec.vni` field to the ClusterPrivateNetwork resource. This
is used by the EVPN/VXLAN feature to bind the Private Network to the
specific VXLAN VNI. The golang type is uint32, but the user configurable
range is 0-(2^24-1).

We also introduce `.status.vni` field that operator-allocated VNI value
will be set. The agent must consume this field instead of consuming
`.spec.vni` directly as the operator performs dynamic allocation when
`.spec.vni` is not set. The operator also performs some checks (like
duplicated VNI check) and only set `.status.vni` when it is valid.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>