cilium-cli: Remove dep. on deprecated CRD in bgp render-api
Remove dependency on the OSS CiliumBGPPeeringPolicy CRD in the
bgp render-api command.

This CRD will be soon removed in OSS, as it has no use when BGPv1
is removed. To allow the render-api command to work after it,
we need to copy the CRD definition into the enterprise subtree
of the cilium-cli, as we use it for unmarshaling the BGPv1
CiliumBGPPeeringPolicy objects in this command.

Other than the CRD itself, we are also copying the related annotations
parsing code, which was used to parse BGP-specific config from
node annotations in BGPv1.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>