bpf: Add tcx_early_hook infra
[ upstream commit a91b33005f47f7e8055bdf7f06e3275cf8e8767e ]

This commit adds a hooking point to the early stage of tcx processing
of inbound traffic from the network. This hook can be used by Cilium
plugins to extend the functionality of tcx. This is equivalent to the
XDP one via commit 30a35be42f6 ("bpf: xdp: add xdp_early_hook infra")
and is needed such that similar extensions can be pulled in when XDP
is not available. Note that similar to 30a35be42f6 this is not a set-
in-stone API and can be removed again in future.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>