fix: Filter by channel for the entry to replace in OLM cat.
When adding a new version, e.g. 1.17.8-cee.1 versions from other
channel than the 1.17 one, e.g. 1.18-beta were considered when
looking for the version to replace. It is now only considering
entires in its own channel.
Note: This does not cover update between channels but it was
not covered earlier either. This would get addressed with #7289.

Signed-off-by: Frederic Giloux <frederic.giloux@isovalent.com>