bgp,rr: Introduce peeringAddressFamily and peerConfigRefV4/V6
Introduce a new concept "peeringAddressFamily" to the ClusterConfig.
This controls how the instance does peering with route-reflector peers.
"IPv4Only" and "IPv6Only" makes either ipv4 or ipv6 peering session and
the "Dual" makes both ipv4 and ipv6 peering sessions.

Since we may have two different peering sessions, we need to associate
different PeerConfig for each session. Two new fields "peerConfigRefV4"
and "peerConfigRefV6" are introduced for this purpose. When the mode is
"IPv4Only" or "IPv6Only", "peerConfigRefV4" or "peerConfigRefV6" should
be used respectively. For "Dual" both can be used.

It is possible to omit peeringAddressFamily. In that case, the default
value will be derived from the Cilium's configuration with the following
rules:

- ipv4.enabled=true && ipv6.enabled=false => IPv4Only
- ipv4.enabled=false && ipv6.enabled=true => IPv6Only
- ipv4.enabled=true && ipv6.enabled=true  => Dual

By combining peeringAddressFamily and the address families enabled with
PeerConfig, users can reflect routes with desired prefix + nexthop
address family combination to the external peer. Below are some concrete
usecases.

- For IPv4-only external peer, use "IPv4Only" and only enable IPv4
  address family with PeerConfig. So that RR only advertises IPv4 prefix
  with IPv4 nexthop.

- For IPv6-only external peer, use "IPv6Only" and only enable IPv6
  address family with PeerConfig. So that RR only advertises IPv6 prefix
  with IPv6 nexthop.

- For dual-stack external peer with extended-nexthop support, use
  "IPv6Only" and enable IPv4 and IPv6 address family with PeerConfig. So
  that RR advertises both IPv4 and IPv6 prefix with IPv6 nexthop.

- For dual-stack external peer without extended-nexthop support, use
  "Dual" and enable IPv4 for IPv4 peering (with peerConfigRefV4) and
  enable IPv6 for IPv6 peering (with peerConfigRefV6).

The existing peerConfigRef field is replaced with peerConfigRefV4/V6.
Fix tests relying on it.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>